backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Flask-Cors==6.0.2
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
PyMySQL==1.1.0
python-dotenv==1.0.0
cryptography==41.0.7
